requirements.txt
python-telegram-bot
google-generativeai
pydub
python-dotenv
ffmpeg-python
